pyproject.toml
[project]
name = "imagestag"
version = "0.1.0"
description = "A fast and efficient image processing and visualization library for Python"
authors = [
    {name = "Michael Ikemann",email = "michael.ikemann@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "nicegui (>=3.4.1,<4.0.0)",
    "pillow (>=10.0.0)",
    "numpy (>=1.26.0)",
    "filetype (>=1.2.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "blessed (>=1.25.0,<2.0.0)",
]

[project.optional-dependencies]
cv = ["opencv-python (>=4.8.0)"]
skimage = ["scikit-image (>=0.22.0)"]
webrtc = ["aiortc (>=1.6.0)"]
pygame = ["pygame (>=2.5.0)"]
kivy = ["kivy (>=2.2.0)"]
dev = ["pytest (>=7.4.0)"]
all = ["opencv-python (>=4.8.0)", "scikit-image (>=0.22.0)", "aiortc (>=1.6.0)", "pygame (>=2.5.0)", "kivy (>=2.2.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
scikit-image = "^0.26.0"
opencv-python = "^4.12.0.88"
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.3.0"
aiortc = "^1.6.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
main_file = "samples/image_demo/main.py"

